두번째 리뷰 반영 - 2단계 자동차경주
