+++ src/main/java/Movable.java
@@ -1,6 +1,6 @@
 //움직일 수 있나? 만 보는 인터페이스
 @FunctionalInterface
 public interface Movable {
-  boolean canMove();
+    boolean canMove();
 }
